LayoutTests/inspector/debugger/setBlackboxBreakpointEvaluations-expected.txt

Tests Debugger.setBlackboxBreakpointEvaluations.


== Running test suite: Debugger.setBlackboxBreakpointEvaluations
-- Running test case: Debugger.setBlackboxBreakpointEvaluations.UnblackboxBreakpointEvaluations.BreakpointCondition.False
Evaluating 'createScripts("UnblackboxBreakpointEvaluations_BreakpointCondition_False")'...
Blackboxing 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Inner.js'...
Setting breakpoint in 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Inner.js'...
Unblackboxing breakpoint evaluations...
Evaluating 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Outer(2)'...
PASS: Should change value of 'x' inside 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Inner.js'
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Inner'.
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointCondition_False_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.UnblackboxBreakpointEvaluations.BreakpointCondition.True
Evaluating 'createScripts("UnblackboxBreakpointEvaluations_BreakpointCondition_True")'...
Blackboxing 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js'...
Setting breakpoint in 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js'...
Unblackboxing breakpoint evaluations...
Evaluating 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Outer(2)'...

PAUSED: 'BlackboxedScript' at 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Outer:3:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "UnblackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js:2:0"
  }
}
Resuming...
PASS: Resumed.
PASS: Should change value of 'x' inside 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js'
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Inner'.
PASS: Should pause in 'UnblackboxBreakpointEvaluations_BreakpointCondition_True_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.UnblackboxBreakpointEvaluations.BreakpointAction.Pause
Evaluating 'createScripts("UnblackboxBreakpointEvaluations_BreakpointAction_Pause")'...
Blackboxing 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js'...
Setting breakpoint in 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js'...
Unblackboxing breakpoint evaluations...
Evaluating 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Outer(2)'...

PAUSED: 'BlackboxedScript' at 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Outer:3:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js:2:0"
  }
}
Resuming...
PASS: Resumed.
PASS: Should change value of 'x' inside 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js'
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Inner'.
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointAction_Pause_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.UnblackboxBreakpointEvaluations.BreakpointAction.AutoContinue
Evaluating 'createScripts("UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue")'...
Blackboxing 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner.js'...
Setting breakpoint in 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner.js'...
Unblackboxing breakpoint evaluations...
Evaluating 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Outer(2)'...
PASS: Should change value of 'x' inside 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner.js'
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner'.
PASS: Should not pause in 'UnblackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.BlackboxBreakpointEvaluations.BreakpointCondition.False
Evaluating 'createScripts("BlackboxBreakpointEvaluations_BreakpointCondition_False")'...
Blackboxing 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Inner.js'...
Setting breakpoint in 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Inner.js'...
Blackboxing breakpoint evaluations...
Evaluating 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Outer(2)'...
PASS: Should change value of 'x' inside 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Outer.js'
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Inner'.
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointCondition_False_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.BlackboxBreakpointEvaluations.BreakpointCondition.True
Evaluating 'createScripts("BlackboxBreakpointEvaluations_BreakpointCondition_True")'...
Blackboxing 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js'...
Setting breakpoint in 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js'...
Blackboxing breakpoint evaluations...
Evaluating 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Outer(2)'...

PAUSED: 'BlackboxedScript' at 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Outer:3:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "BlackboxBreakpointEvaluations_BreakpointCondition_True_Inner.js:2:0"
  }
}
Resuming...
PASS: Resumed.
PASS: Should change value of 'x' inside 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Outer.js'
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Inner'.
PASS: Should pause in 'BlackboxBreakpointEvaluations_BreakpointCondition_True_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.BlackboxBreakpointEvaluations.BreakpointAction.Pause
Evaluating 'createScripts("BlackboxBreakpointEvaluations_BreakpointAction_Pause")'...
Blackboxing 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js'...
Setting breakpoint in 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js'...
Blackboxing breakpoint evaluations...
Evaluating 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Outer(2)'...

PAUSED: 'BlackboxedScript' at 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Outer:3:4'.
{
  "originalReason": "Breakpoint",
  "originalData": {
    "breakpointId": "BlackboxBreakpointEvaluations_BreakpointAction_Pause_Inner.js:2:0"
  }
}
Resuming...
PASS: Resumed.
PASS: Should change value of 'x' inside 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Outer.js'
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Inner'.
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointAction_Pause_Outer'.

-- Running test case: Debugger.setBlackboxBreakpointEvaluations.BlackboxBreakpointEvaluations.BreakpointAction.AutoContinue
Evaluating 'createScripts("BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue")'...
Blackboxing 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner.js'...
Setting breakpoint in 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner.js'...
Blackboxing breakpoint evaluations...
Evaluating 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Outer(2)'...
PASS: Should change value of 'x' inside 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Outer.js'
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Inner'.
PASS: Should not pause in 'BlackboxBreakpointEvaluations_BreakpointAction_AutoContinue_Outer'.
